Update README.md
This is my 3rd project made it using html and css.I'm thrilled to share my latest creation with you all. This profile card is a perfect blend of HTML and CSS.I have created it with help the of one of the YouTube creator "CodingLab" 
I'm always looking for opportunities to grow and collaborate, so if you have any feedback, questions, or potential projects, feel free to reach out. 🤝